refactor: extract tailing logic into tail service
